+++ federation/queue.ts
@@ -3,6 +3,8 @@ export interface SenderKeyJwkPair {
   privateKey: JsonWebKey;
 }
 
+export type Message = OutboxMessage | InboxMessage;
+
 export interface OutboxMessage {
   type: "outbox";
   keys: SenderKeyJwkPair[];
@@ -10,4 +12,11 @@ export interface OutboxMessage {
   inbox: string;
   trial: number;
   headers: Record<string, string>;
+}
+
+export interface InboxMessage {
+  type: "inbox";
+  baseUrl: string;
+  activity: unknown;
+  handle: string | null;
 }